requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
Werkzeug
